# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce Core API Documentation",
        "description": "API documentation for Laravel E-Commerce Core project including Auth, Products, Categories, Cart, Orders, Coupons, Reviews/Comments, Addresses, and Admin Management.",
        "contact": {
            "name": "API Support Team",
            "email": "admin@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lấy danh sách danh mục sản phẩm",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Danh sách danh mục"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Tạo danh mục mới (Admin)",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Thời trang nam"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mô tả danh mục..."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo danh mục thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Chi tiết danh mục",
                "operationId": "ff3ab984d3fe877eb16da6009394d694",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết danh mục"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Cập nhật danh mục (Admin)",
                "operationId": "3146e96866d62a90bd01a7070fd61470",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Thời trang nữ"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mô tả danh mục..."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Xóa danh mục (Admin)",
                "operationId": "88c22abd8149aea4b3bdcb280a0fddf7",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/coupons/validate": {
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Kiểm tra mã giảm giá",
                "operationId": "c4b7f985e96587722f329a0e39b1fcd3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "order_amount"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "WELCOME10"
                                    },
                                    "order_amount": {
                                        "type": "number",
                                        "example": 500000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mã giảm giá hợp lệ"
                    },
                    "422": {
                        "description": "Mã giảm giá không hợp lệ hoặc không đủ điều kiện"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {your_token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}